Split the Relay Control and the Input Reader
